delete file is note is deleted

+++ src/main/java/com/intellinote/note/NoteController.java
@@ -101,6 +101,9 @@ public class NoteController {
     
     @DeleteMapping("/auth/users/{username}/notes/{id}")
     public @ResponseBody void deleteNote(@PathVariable int id){
-        nr.deleteById(id);
+        Note n = nr.getOne(id);
+        if(ss.removeFile(n.getPath())){
+            nr.deleteById(id);
+        }
     }
 }
